net-restapi/net-restapi/requirements.txt
Django==4.2.10
djangorestframework==3.14.0
djangorestframework-simplejwt[crypto]==5.3.1
pillow==10.3.0
drf-spectacular[sidecar]==0.27.2